+++ CHANGELOG.md
@@ -10,7 +10,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ### Added
 
-- Command `add-server` to add server from command line.
+- Command `server add` to add server from command line.
 
 ## [0.1.11] - 2023-10-31